refactor(trading-signals): promote linear regression to a documented utility
calculateLinearRegression becomes getLinearRegression under src/util/,
matching the get* naming of its siblings, with a co-located test. The
README utility list now covers the full exported surface: linear
regression, percentage change, share, sliding window update, and typical
price were exported but undocumented.
